requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.1.8
django-cors-headers==3.14.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastapi==0.82.0
fastjsonschema==2.16.2
filelock==3.7.1
Flask==2.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
ipykernel==6.15.1
ipython==8.13.1
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.16.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==2.0.4
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.5
notebook==6.5.2
notebook_shim==0.2.0
numpy==1.24.3
openpyxl==3.0.10
packaging==21.3
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.11
pickleshare==0.7.5
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.12.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==23.2.0
requests==2.28.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
starlette==0.19.1
terminado==0.17.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.12
uvicorn==0.18.3
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.1.2
